feat: Add Remove-AdobeUser command to remove Adobe user
The Remove-AdobeUser command is added to remove a user from Adobe. It takes a mandatory parameter for the email address of the user and an optional parameter to specify whether to delete the user account. The command checks if the user is connected to Adobe using the Connect-Adobe command. It then constructs the necessary data and sends a request to remove the user from the organization. The testOnly query parameter is used to determine if the command should be executed or just tested.