feat: Add New-AdobeUser function to create a new Adobe user
The New-AdobeUser function is added to create a new user in Adobe. It takes parameters for the user's email address, country, first name, last name, option, and type. The function checks if the user is connected to Adobe using the Connect-Adobe command. It then constructs the necessary data and sends a request to create the user in the organization. The testOnly query parameter is used to determine if the function should be executed or just tested.