feat: Add Set-AdobeGroup function to update Adobe user group
The Set-AdobeGroup function is added to update an existing Adobe user group. It takes parameters for the group's name, new name, and description. The function checks if the user is connected to Adobe using the Connect-Adobe command. It then constructs the necessary data and sends a request to update the group in the organization. The testOnly query parameter is used to determine if the function should be executed or just tested.